+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   sptotal-vignette: sptotal-vignette.html
-last_built: 2023-03-17T11:54Z
+last_built: 2023-03-20T16:27Z
 
