### 2023-04-11 #5283 @fabian-s
:wave: @trashbirdecology & @garretrc

Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
