### 2023-04-11 highamm/sptotal#7 @fabian-s
see also [joss-reviews/#5283](https://github.com/openjournals/joss-reviews/issues/5283)
