+++ inst/paper/paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={e0282524},
   year={2023},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0282524}
 }
 
 @article{pebesma2018simple,
@@ -16,7 +17,8 @@
   volume={10},
   number={1},
   pages={439},
-  year={2018}
+  year={2018},
+  doi={10.32614/rj-2018-009}
 }
 
 @article{ver2002sampling,
@@ -27,7 +29,8 @@
   number={2},
   pages={152--161},
   year={2002},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/11956860.2002.11682701}
 }
 
 @article{dumelle2022comparison,
@@ -38,7 +41,8 @@
   number={9},
   pages={2018--2029},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/2041-210X.13919}
 }
 
 @article{ver2013comparison,
@@ -49,7 +53,8 @@
   number={3},
   pages={e59129},
   year={2013},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0059129}
 }
 
 @book{fix1985discriminatory,
@@ -57,7 +62,8 @@
   author={Fix, Evelyn},
   volume={1},
   year={1985},
-  publisher={USAF school of Aviation Medicine}
+  publisher={USAF school of Aviation Medicine},
+  doi={10.1037/e471672008-001}
 }
 
 @article{stevens2004spatially,
@@ -78,7 +84,8 @@
 	journal = {Environmental and Ecological Statistics},
 	author = {Ver Hoef, Jay M},
 	year = {2008},
-	pages = {3--13}
+	pages = {3--13},
+	doi = {10.1007/s10651-007-0035-y}
 }
 
 @book{cressie2015statistics,
@@ -96,7 +103,8 @@
   number={7},
   pages={683--691},
   year={2004},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cageo.2004.03.012}
 }
 
 @Manual{ribeiro2007geor,
