+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   sptotal-vignette: sptotal-vignette.html
-last_built: 2023-03-20T16:27Z
+last_built: 2023-04-11T17:11Z
 
