### 2023-04-11 #5283 @highamm
Thank you @kthyng and @fabian-s for your help with getting this started.

@fabian-s I have added the missing doi's and I will get started working improving the testing coverage and replacing the 1:n_iter iterators mentioned in highamm/sptotal#7

Thank you both again for your time!
