### 2023-04-12 #5283 @fabian-s
wave @potterzot @Athene-ai @garretrc

Would any of you be willing to review this submission for JOSS?
