### 2023-04-12 #5283 @Athene-ai
Hi!

Yes. I am interested

<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
Privo
di virus.www.avast.com
<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
<#DAB4FAD8-2DD7-40BB-A1B8-4E2AA1F9FDF2>

On Wed, 12 Apr 2023 at 11:00, Fabian Scheipl ***@***.***>
wrote:

> wave @potterzot <https://github.com/potterzot> @Athene-ai
> <https://github.com/Athene-ai> @garretrc <https://github.com/garretrc>
>
> Would any of you be willing to review this submission for JOSS?
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
