### 2023-04-12 #5283 @kthyng
@openjournals/dev Could you take a look at this? We seem to be caught in a loop where we can't properly assign anyone.
