### 2023-04-12 d70ecba
started updating tests for all functions and started fixing the readability of some of the documentation.
