### 2023-04-12 f32c99e
added more tests and edited more of the documentation to make the documentation more readable. Also removed aes_() function from plotting functions, as it has be deprecated in the newest version of ggplot2.
