+++ tests/testthat/test-sv.R
@@ -1,0 +1,5 @@
+test_that("semi-variogram is generated", {
+  expect_error(sv(na.omit(exampledataset),
+                  "xcoords", "ycoords", "counts"),
+               NA)
+})
