+++ R/slmfit.R
@@ -4,24 +4,35 @@
 #' parameters, given spatial coordinates and a model formula.
 #'
 #' @param formula is an \code{R} linear model formula specifying the
-#' response variable as well as covariates for predicting the response on the unsampled sites.
-#' @param data is a data frame or tibble with the response column, the covariates to
-#' be used for the block kriging, and the spatial coordinates for all of the sites. Alternatively, data can be an \code{sp} Spatial Points Data Frame or \code{sf} object with POINT geometry.
-#' @param xcoordcol is the name of the column in the data frame with x coordinates or longitudinal coordinates
-#' @param ycoordcol is the name of the column in the data frame with y coordinates or latitudinal coordinates
-#' @param areacol is the name of the column with the areas of the sites. By default, we assume that all sites have equal area, in which
+#' response variable as well as covariates for predicting the
+#' response on the unsampled sites.
+#' @param data is a data frame or tibble with the response column,
+#' the covariates to be used for the block kriging, and the
+#' spatial coordinates for all of the sites. Alternatively, data can be
+#' an \code{sp} Spatial Points Data Frame or \code{sf} object with
+#' POINT geometry.
+#' @param xcoordcol is the name of the column in the data frame with
+#' x coordinates or longitudinal coordinates
+#' @param ycoordcol is the name of the column in the data frame with
+#' y coordinates or latitudinal coordinates
+#' @param areacol is the name of the column with the areas of the sites.
+#' By default, we assume that all sites have equal area, in which
 #' case a vector of 1's is used as the areas.
 #' @param stratacol is the name of the the column with the
 #' stratification variable, if strata are to be fit separately,
 #' with different covariance parameter estimates.
-#' @param CorModel is the covariance structure. By default, \code{CorModel} is
-#' Exponential but other options include the Spherical and Gaussian.
+#' @param CorModel is the covariance structure. By default,
+#' \code{CorModel} is Exponential but other options include the
+#' Spherical and Gaussian.
 #' @param estmethod is either the default \code{"REML"} for restricted
 #' maximum likelihood to estimate the covariance parameters and
 #' regression coefficients or \code{"ML"} to estimate the covariance
 #' parameters and regression coefficients. This argument can also be set to
 #' \code{"None"}, in which case \code{covestimates} must be provided.
-#' @param covestimates is an optional vector of covariance parameter estimates (nugget, partial sill, range). If these are given and \code{estmethod = "None"}, the the provided vector are treated as the estimators to create the covariance structure.
+#' @param covestimates is an optional vector of covariance
+#' parameter estimates (nugget, partial sill, range). If these are
+#' given and \code{estmethod = "None"}, the the provided vector are
+#' treated as the estimators to create the covariance structure.
 #' @return a list of class \code{slmfit} with \itemize{
 #'   \item the spatial covariance estimates
 #'   \item the regression coefficient estimates
@@ -53,7 +64,8 @@
 #'
 #' data(exampledataset) ## load a toy data set
 #' exampledataset$strata <- c(rep("A", 19), rep("B", 21))
-#' strataobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset, stratacol = "strata",
+#' strataobj <- slmfit(formula = counts ~ pred1 + pred2,
+#'   data = exampledataset, stratacol = "strata",
 #' xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
 #' summary(strataobj)
 #' @import stats
@@ -123,7 +135,12 @@ slmfit <- function(formula, data, xcoordcol, ycoordcol,
     colnames(datapredsonly) <- all.vars(formula)[-1]
 
[…]
