### 2023-04-13 #5283 @xuanxu
I've found the problem, editorialbot does not recognize `fabian-s` as a GitHub user (is missing the `@`)
Running `@editorialbot add @fabian-s as editor` should fix it.
