### 2023-04-13 #5283 @fabian-s
> I've found the problem, editorialbot does not recognize `fabian-s` as a GitHub user (is currently listed in the issue body missing the `@`) Running `@editorialbot add @fabian-s as editor` should fix it.

oh man.. sorry about that, i'm an idiot :see_no_evil:
