### 2023-04-13 #5363 @editorialbot
[REVIEW]: sptotal: an R package for predicting totals and weighted sums from spatial data
**Submitting author:** @highamm (<a href="http://orcid.org/0009-0006-4217-625X">Matt Higham</a>)
**Repository:** https://github.com/highamm/sptotal
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.1
