### 2023-04-13 #5363 @fabian-s
see also https://github.com/highamm/sptotal/issues/7
