### 2023-04-13 highamm/sptotal#7 @fabian-s
see also https://github.com/openjournals/joss-reviews/issues/5363
