### 2023-04-13 highamm/sptotal#7 @highamm
Hello! Thanks again for getting this started. I just have a quick update that we've added a bit more testing with the newer major release of testthat. We have also changed the 1:n_iter and sapply() code in the latest push. There are still some tests we plan on adding this week to address some of the lines of code still not covered, but we have made some progress. Below is the output from `goodpractice::gp(".")` on the newest version.

```
── GP sptotal ────────────────────────────────────────────────────────────────

It is good practice to

    R/plot.slmfit.R:33:NA
    R/predict.slmfit.R:53:NA
    R/predict.slmfit.R:54:NA
    R/predict.slmfit.R:71:NA
    R/predict.slmfit.R:72:NA
    ... and 94 more lines

    R/data.R:109:81
    R/data.R:111:81
    R/data.R:112:81
    R/data.R:113:81

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
```
