### 2023-04-13 highamm/sptotal#7 @fabian-s
great, thx!

89% coverage seems fine to me already, long code lines are irrelevant.

just checked your imports: if you adapt https://github.com/highamm/sptotal/blob/7b5ecd9ae733ae62c6d6bb0a8eba13fe32418c20/R/print.summary.slmfit.R#L23, https://github.com/highamm/sptotal/blob/7b5ecd9ae733ae62c6d6bb0a8eba13fe32418c20/R/print.summary.slmfit.R#L74 to use @importFrom and pick the specific stats-functions you need this is perfect as well.

EDIT:
importing all of ggplot2 and graphics seems fine to me.
there are some more lines where you import all of stats in other roxygen-blocks, see https://github.com/highamm/sptotal/search?q=import
