### 2023-04-13 3c914c3
cleaned up imports from stats and ggplot2, moved graphics to suggests, and removed viridis as an import (which now comes bundles with ggplot2).
