+++ R/GR2.R
@@ -2,7 +2,6 @@
 #'
 #' @param object is an object of class \code{slmfit}.
 #' @return the value for generalized R-squared
-#' @import stats
 #' @examples
 #' data(exampledataset) ## load a toy data set
 #' slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
