+++ R/predict.slmfit.R
@@ -40,7 +40,6 @@
 #' slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
 #' xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
 #' predict(slmobj)
-#' @import stats
 #' @export
 
 
