+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   sptotal-vignette: sptotal-vignette.html
-last_built: 2023-04-13T12:25Z
+last_built: 2023-04-13T18:26Z
 
