### 2023-04-13 highamm/sptotal#7 @highamm
Sounds good! The most recent commit [3c914c3](https://github.com/highamm/sptotal/commit/3c914c3d763b9b0882d67655a3d7a9f24c870864) fixed these import issues (and addresses the `ggplot2` and `graphics` imports: looking at them again helped me catch the fact that we no longer need to import anything from `viridis`, whose main functions now come bundled with `ggplot2`).

Thank you again for the suggestion!
