### 2023-05-09 #5363 @garretrc
H

> @garretrc: could you please fill the remainder of the checklist and add comments or open issues for the remaining open points in your checklist so we can discuss what steps @highamm needs to take to get this paper published?

Yes, I'll get to work this week, unfortunately up against a submission deadline today. From what I've read and messed with, I think I will have some helpful feedback!
