### 2023-05-16 447e129
added a few more function tests.
