+++ tests/testthat/test-strata.R
@@ -12,19 +12,21 @@ test_that("Stratification fits", {
   expect_error(stratamod, NA)
 })
 
-stratamod <- stratafit(counts ~ pred1, data = exampledataset,
-                       xcoordcol = "xcoords", ycoordcol = "ycoords",
-                       stratacol = "stratavar")
 
-test_that("Stratification summary output does not change", {
+
+test_that("including stratacol in slmfit analyzes strata separately", {
+  stratamod <- stratafit(counts ~ pred1, data = exampledataset,
+                         xcoordcol = "xcoords", ycoordcol = "ycoords",
+                         stratacol = "stratavar")
+  expect_error(stratamod, NA)
+  expect_length(stratamod, 3)
+  expect_error(print(stratamod), NA)
   expect_snapshot(summary(stratamod))
 })
 
-test_that("stratafit objects are printed", {
-  expect_error(print(stratamod), NA)
-})
-
-
+stratamod <- stratafit(counts ~ pred1, data = exampledataset,
+                       xcoordcol = "xcoords", ycoordcol = "ycoords",
+                       stratacol = "stratavar")
 
 test_that("stratamod has length equal to the number of strata", {
   expect_equal(length(stratamod), nlevels(exampledataset$stratavar),
@@ -46,6 +48,7 @@ test_that("helper functions can be used on stratafit objects", {
   expect_equal(length(fitted(stratamod[[2]])), 15)
   expect_equal(AIC(stratamod[[3]]), 93.6, tolerance = 0.1)
 })
+
 
 
 test_that("including strata in stratacol and as fixed effect generates error", {
