+++ tests/testthat/test-sim.R
@@ -1,0 +1,3 @@
+test_that("Points can be simulated on a grid", {
+  expect_error(pointSimSyst(), NA)
+})
