### 2023-05-23 #5363 @highamm
@fabian-s Great, thank you for the update!

The tagged release version of the software is v1.0.1 given at <https://github.com/highamm/sptotal/tree/v1.0.1>

The software is now archived in Zenodo with DOI `10.5281/zenodo.7962631` at <https://doi.org/10.5281/zenodo.7962631>. The authors and title do match the authors and title in the paper and on the software.

The DOI for the archived version is `10.5281/zenodo.7962631`

Let me know if I missed anything here!
